Handle and display errors from ansible runs
Implemented support for handling error codes and stderr output from ansible
runs. Also included is the support for retries (default is 3) when a run
fails and a new status `provisioning_failure'  on OpenEdXInstance to mark
such failed deployments.

The UI was also updated to display prominently the last error happened on
instances with the status `provisioning_failure'.